test: Add tests for the TeamPermission
